experiment with i2c clock frequency
Investigate kernel and i2c handling.
This revealed a problem with pull up values.
The pull up values should be changed to 4.7K or less.
Also revealed issues with the stop bit, SCL low timout,
and rtc counter synchronization.
This should be changed back to BITRATE_FAST when
the pulls ups and changed.