add rtc synchronization
Used for experimenting with the RTC during sleep
Specifically, how RCONT and COUNT are updated
during and after sleep.
It appears that there can be a significant delay
after sleep before COUNT is updated.  This can cause
problems for k_set_timeout() since the value returned
by COUNT can be inaccurate and a timeout could be set that
has already past.
It appears the only reliable way to ensure COUNT reads reliably
is to use a manual RREQ.