wip:  support standby sleep state
This is the first working version of the driver that
supports the standby sleep state.
Reference is made to the SAMD21 datasheet
26.6.3.7 and 26.6.5
The OSC8M should be used at the source oscillator
for the GCLK used to clock the sercom.  This oscillator
is recommended because is has a fast wake time.
The RUNSTDBY and ONDEMAND bits must be set
for the oscillator so that any data in the transmit buffers
is sent if the device is put to sleep.  These bits also ensure
the sercom will receive data when using the start of frame
detection.