drivers: spi_nand: change timeout delays
I was observing that the nand chip frequenty timed out
during device operations (read, write and erase).

This commit changes the timeouts an order of magnitude
to 100ms each.

Initial observation so far is that timeouts are infrequent now.
I have not observed a single timeout in several days of
testing.