drivers: murata-lora: remove cmd complete callback
The problem with start of frame detection waking the mcu
has been resolved.  Now the mcu will wake every time a
character is received from the modem.

With this working, it has been shown that it is not necessary
to stay awake while receiving the data.  The interrupt fires
fast enough to receive the data from the modem
at 19200 baud.  I do not think the module can support
faster baud rates.