wip:  resolve boot hang
The software could hang coming out of a reset.
A problem was identified with the amount of time
after a pin reset.  If this is too short, the i2c port
does not respond and the initialization is not completed
properly.

Also, we should try to detect the device before creating
the message thread since the message thread reads
the data register.  So, the i2c test was moved before
the creation of the message thread.

Add shell command to set pvt rate.