wip: automate pvt solution poll
Add shell commands to save and load BBR configurations.
Add infrastructure to handle timepulse.  This is not yet complete.
Use k_poll and signals to control trigger thread.

Requesting a pvt solution requires waking the module
and then waiting for a fix.  A PVT trigger was added to
coordinate these activities concurrently.  The application
now requests the pvt solution 10 seconds before it is
required to allow for the nav acquisition phase.