wip: improve battery disconnect detection
Detecting a battery disconnect is unreliable, particulary
when the battery voltage is very low.
Some code / ideas from the particle-iot power manager were reviewed.
This commit implements a method to count
the number of charge cycles which is a more reliable
way to determine if the battery has been disconnected.
Added shell framework and initial commands.