wip: poll the connect status
It may be helpful to be able to tell the driver to reconfigure
the pmic / charge settings.  We will need to investigate the settings
framwork to determine how best to do this.  In the mean
time, use signals so that we can tell the pmic process
thread what to do.

Use polling method to determine battery connection
status.
Use k_poll and signals to control state machine updates instead
of semaphore.
Clean up.