isb_defconfg:  use k_cpu_idle()
k_cpu_idle()  idles the processor.
This commit changes the devicetree config so that we use
k_cpu_idle() instead of custom code for power management.
This simplifies the amount of custom code that needs to be supported
for the application.

To go into deep sleep (STANDBY), the PM / SCB must be configured
to enable STANDBY mode when WFI (this is done in the app).

For the SAMD21, the RTC counter stops continously updating
after waking from STANDBY.  So, we need to enable
SYS_POWER_MANAGEMENT so that our we can restart the
RTC counter.  This is done with a call to rtc_wake() from the
z_clock_idle_exit() function.  These functions are part of the
sam0_rtc_timer driver in the out-of-tree repository.

The SAMD51 does not require a call to z_clock_idle_exit()
after wake up to restart the RTC counter so we can compile
the SAMD51 without SYS_POWER_MANAGEMENT.