defconfig:  fix HW_CYCLES_PER_SEC
HW_CYCLES_PER_SEC is not set correctly.  There was a problem
with cycle count calculations and the old "fix" was to change
the cycles per sec.

We found the problem in z_timer_cycle_get_32().  This fucntion
incorrectly returned the number of RTC HW cycles.  The kernel
requires that the number of TICKS be returned so that it can
properly calculate the number of SYS HW cycles.

This commit converts the RTC HW cycles returned from the
rtc_count() function to TICKS.