drivers: sam0_rtc_timer:  COUNTSYNC errata
Address COUNTSYNC errata for SAMD5x devices
When CTRLA.COUNTSYNC is enabled, the first COUNT value is not
correctly synchronized and thus it is a wrong value.

Workaround:
After enabling COUNTSYNC, read the COUNT register until its value
is changed when compared to its first value read. After this, all
consequent value read from the COUNT register is valid.

This also ensures that RCONT is properly set and RTC is counting
when using SAMD2x.