spi_nand:  refactor status wait
The device datasheet is not clear when the ECC bits
are set and reset.  When checking the ECC bits in the wait loop,
the file system does not mount.

So, this commit disables ECC bit checking in
spi_nand_wait_until_ready() and re-orders the checking of
the PROGF, ERASEF and OIP bits.  Now, OIP is checked last so that
we exit early if there is a PROGF or ERASEF so that we do
not miss these states when we return from wait.