spi_nand: update DTS page read time
Page reads can take longer than the max time specified in the datasheet.
It is not clear why this is happening.  So, the spi_nand driver has been
updated to try the read again if it times out.  This seems to work
well during testing compared to simply extending the duration
of the timeout.

Even with long timeout durations we would still observe periodic
time outs.  Using the retry method, we have not oberved a timeout
after the retry.

TODO:
We shoudl consider if the same technique should be used for the
page prog time as well.  That is, if we time out during a page write,
we should retry the page prog to avoid arbitrarily long prog times.