murata:  disable confirmed uplinks
If an ack is not received within the rx window of the uplink
that requested the ack then the ack either was not sent from
the server or it was not received by the end device.

Note that the ACK can be lost for any number of reasons
including a poor radio environment.  Chirpstack strictly
enforces the LoRaWAN requirement that there cannot be
multiple ACK's to an uplink.  So, the consequence of
loosing an ACK is the retries are ignored by Chirpstack.

So, we either ignore the fact we did not receive the ack or we
must wait until LoRaMac is not busy (no longer retransmitting)
and send another confirmed uplink.  This makes confirmed
uplinks dubious.

So we ignore that we did not receive the ack.

Note that confirmed uplinks cause undesirable side effects.
In particular, an uplink that follows the confirmed uplink
will have its ADR changed.  In my observations, for some reason
the ADR appears to be reset to DR0 which has a very limited
payload size (11 bytes).

The consequence is that it is likely that the uplink following
the confirmed uplink will have it's payload removed because
the payload is too large for the DR0 data rate.

Set lora.status.req_ack to true to turn on confirmed uplinks if you
really want to use them.  Confirmed uplinks will occur with a period
set by LORA_HEARTBEAT_TIMEOUT.