murata:  add documentation
Add reference to changing data rate when the ACK retransmission
process is executed.

This reference documents why the data rate is changing to DR0
when an ACK is lost.  When this happens, the max payload is
11 bytes.  If the payload is larger, then SendFrame() strips the
payload and sends the frame to "flush the MAC commands".

Unfortunately, when this occurs, the application is not notified
that the payload was stripped.