spi_nand:  add parameter timers
There appears to be an issue about how to interpret the "maximum"
timing values provided in the datasheet for spi nand devices.
Empircally, we have found that the maximum values to page
read, page program and block erase provided in the datasheet are
not guaranteed.

Therefore, this commit adds timers to track cummulative and
average time for key nand performance parameters so we can
better evaluate the actual performance of the nand device:
page read time
page program time
block erase time