spi_nand:  refactor operation timer
Use delta timer to ensure timeout compare works properly when the cycle
timer wraps around.

Also fixes an issue where if we timed out during page read we
re-issued the page read command.  Rather, we should continue to wait
for the operation to complete.  This problem was identified
when we  observed frequent page read timeouts suggesting the
maximum value provided in the datasheet and parameter page is
not guaranteed.