spi_nand:  extend timeouts
The maximum times provided in the device datasheets for
page read time
page program time
block erase time

do not appear to be guaranteed.  It has been observed that frequent
timeouts occur when using the devices.  To alleviate the problem,
we extend the timeouts to be a multiple of the maximum values
stated in the datasheet.

In another commit, we added support for tracking the actual
cummulative times and counts so we can determine an average for
these critical timing parameters.