witap_log:  fix file handle and sync
Add function to enable periodic file sync to flash.

Care needs to be taken to ensure the file write and file
close operations occur only when the file is open and
there is an valid file handle.
Similarly, a data file should only be opened when it is
not already open.